requirements.txt
# Core AI and ML libraries
openai==1.3.5
anthropic==0.7.7
langchain==0.1.0
langchain-openai==0.0.2

# Voice processing
speech-recognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
wave==0.0.2

# Web automation and scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0

# GUI and interface
tkinter
customtkinter==5.2.0
pygame==2.5.2

# Data processing
pandas==2.1.3
numpy==1.24.3
json5==0.9.14

# Database
sqlite3
sqlalchemy==2.0.23

# API integrations
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1

# Utility libraries
python-dotenv==1.0.0
schedule==1.2.0
datetime
asyncio
threading
logging
configparser

# Additional utilities
geopy==2.4.1
python-dateutil==2.8.2
regex==2023.10.3